added head, header, and footer commands
