+++ lib/livingstyleguide/filters.rb
@@ -32,4 +32,4 @@ require 'livingstyleguide/filters/colors'
 require 'livingstyleguide/filters/css'
 require 'livingstyleguide/filters/style'
 require 'livingstyleguide/filters/data'
-require 'livingstyleguide/filters/html_head'
+require 'livingstyleguide/filters/layout'
